+++ benchmark/actors/singlethreaded/define_supervisor.hpp
@@ -48,9 +48,9 @@ public:
         } else if (actor_0_) {
             auto [needs_sched, future] = actor_zeta::send(actor_0_.get(), &Actor::start);
             actor_zeta::detail::ignore_unused(future);
-            actor_0_->resume(1);
-            actor_1_->resume(1);
-            actor_0_->resume(1);
+            (void)actor_0_->resume(1);
+            (void)actor_1_->resume(1);
+            (void)actor_0_->resume(1);
         }
         co_return;
     }